+++ Synth/Noise/INoiseGenerator.cs
@@ -1,9 +1,6 @@
-﻿using W = System.Func<double, double, double, byte>;
+﻿using W = System.Func<double, double, double, short>;
 
 namespace Synth.Noise
 {
-    public interface INoiseGenerator
-    {
-        W Next();
-    }
+    public interface INoiseGenerator { W Next(); }
 }
